requirements/apps.txt
-e svn+http://code.djangoproject.com/svn/django/branches/releases/1.2.X#egg=django
django-mptt